# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dota 2 Auction Draft API",
    "description": "REST API for the Dota 2 Auction Draft platform. Manages competitions, players, captains, auctions, tournaments, and user accounts.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "QueuePool": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "enabled": { "type": "boolean" },
          "min_mmr": { "type": "integer" },
          "max_mmr": { "type": "integer" },
          "pick_timer": { "type": "integer" },
          "best_of": { "type": "integer" },
          "team_size": { "type": "integer" },
          "lobby_server_region": { "type": "integer" },
          "lobby_game_mode": { "type": "integer" },
          "lobby_league_id": { "type": "integer" },
          "lobby_dotv_delay": { "type": "integer" },
          "lobby_cheats": { "type": "boolean" },
          "lobby_allow_spectating": { "type": "boolean" },
          "lobby_pause_setting": { "type": "integer", "description": "0 = unlimited, 1 = limited, 2 = disabled" },
          "lobby_selection_priority": { "type": "integer", "description": "0 = manual, 1 = automatic (coin toss)" },
          "lobby_cm_pick": { "type": "integer", "description": "0 = random, 1 = radiant, 2 = dire" },
          "lobby_auto_assign_teams": { "type": "boolean", "description": "Bot enforces team assignments" },
          "lobby_penalty_radiant": { "type": "integer", "description": "Radiant penalty in seconds" },
          "lobby_penalty_dire": { "type": "integer", "description": "Dire penalty in seconds" },
          "lobby_series_type": { "type": "integer" },
          "lobby_timeout_minutes": { "type": "integer" },
          "xp_win": { "type": "integer" },
          "xp_participate": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "queue_count": { "type": "integer", "description": "Number of players currently waiting in this pool. Included on the public /api/queue/pools listing for pool-selector badges; not persisted in the DB." }
        }
      },
      "QueuePoolInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "enabled": { "type": "boolean", "default": true },
          "min_mmr": { "type": "integer", "default": 0 },
          "max_mmr": { "type": "integer", "default": 0 },
          "pick_timer": { "type": "integer", "default": 30 },
          "best_of": { "type": "integer", "default": 1 },
          "team_size": { "type": "integer", "default": 5 },
          "lobby_server_region": { "type": "integer", "default": 3 },
          "lobby_game_mode": { "type": "integer", "default": 2 },
          "lobby_league_id": { "type": "integer", "default": 0 },
          "lobby_dotv_delay": { "type": "integer", "default": 1 },
          "lobby_cheats": { "type": "boolean", "default": false },
          "lobby_allow_spectating": { "type": "boolean", "default": true },
          "lobby_pause_setting": { "type": "integer", "default": 0 },
          "lobby_selection_priority": { "type": "integer", "default": 0 },
          "lobby_cm_pick": { "type": "integer", "default": 0 },
          "lobby_auto_assign_teams": { "type": "boolean", "default": true },
          "lobby_penalty_radiant": { "type": "integer", "default": 0 },
          "lobby_penalty_dire": { "type": "integer", "default": 0 },
          "lobby_series_type": { "type": "integer", "default": 0 },
          "lobby_timeout_minutes": { "type": "integer", "default": 10 },
          "xp_win": { "type": "integer", "default": 15 },
          "xp_participate": { "type": "integer", "default": 5 }
        }
      }
    }
  },
  "paths": {
...
    "/api/queue/pools/{poolId}": {
      "get": {
        "tags": ["Queue"],
        "summary": "Get queue pool details",
        "parameters": [{ "name": "poolId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Queue pool", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueuePool" } } } },
          "404": { "description": "Pool not found" }
        }
      }
    },
    "/api/queue/players/stats": {
      "get": {
        "tags": ["Queue"],
        "summary": "Per-player W/L over the last N completed queue matches in a pool",
        "description": "Returns win/loss counts computed from each player's most recent completed queue matches in the given pool. Used by the in-queue roster to show a mini form guide.",
        "parameters": [
          { "name": "poolId", "in": "query", "required": true, "schema": { "type": "integer" }, "description": "Queue pool id" },
          { "name": "playerIds", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Comma-separated player ids" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 50 }, "description": "How many recent completed matches per player to aggregate over" }
        ],
        "responses": {
          "200": {
            "description": "Array of per-player W/L; players with no completed matches in the pool are omitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "playerId": { "type": "integer" },
                      "wins": { "type": "integer" },
                      "losses": { "type": "integer" }
                    },
                    "required": ["playerId", "wins", "losses"]
                  }
                }
              }
            }
          },
          "400": { "description": "poolId missing" }
        }
      }
    },
    "/api/admin/queue/pools": {
      "get": {
        "tags": ["Queue"],
        "summary": "Admin: list all queue pools",
        "description": "Returns all queue pools including disabled ones. Requires manage_queue_pools permission.",
        "responses": { "200": { "description": "Array of all queue pools", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/QueuePool" } } } } } }
      },
      "post": {
        "tags": ["Queue"],
        "summary": "Admin: create queue pool",
        "description": "Creates a new queue pool. Requires manage_queue_pools permission.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueuePoolInput" } } } },
        "responses": {
          "201": { "description": "Created pool", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueuePool" } } } },
          "400": { "description": "Name is required" }
        }
      }
    },
    "/api/admin/queue/pools/{id}": {
      "put": {
        "tags": ["Queue"],
        "summary": "Admin: update queue pool",
        "description": "Updates an existing queue pool. Requires manage_queue_pools permission. Only provided fields are updated.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueuePoolInput" } } } },
        "responses": {
          "200": { "description": "Updated pool", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueuePool" } } } },
          "404": { "description": "Pool not found" }
        }
      },
      "delete": {
        "tags": ["Queue"],
        "summary": "Admin: delete queue pool",
        "description": "Deletes a queue pool and cascades to queue_matches. Requires manage_queue_pools permission.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
    "/api/admin/queue/matches/{id}/cancel": {
      "post": {
        "tags": ["Queue"],
        "summary": "Admin: cancel a queue match",
        "description": "Cancels an active queue match. Notifies all players via socket, cleans up in-memory state, and sets DB status to cancelled. Refuses once the latest match_lobbies row is 'completed' (bot has left the lobby after game_started) — at that point the Dota match is running on its own. Requires manage_queue_pools permission.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Match cancelled" },
          "409": { "description": "Match already started — bot has left the lobby" }
        }
      }
    },
...
    "/api/admin/queue/kick/{playerId}": {
      "post": {
        "tags": ["Queue"],
        "summary": "Admin: kick a player out of their current queue",
        "parameters": [{ "name": "playerId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "reason": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Kick result" } }
      }
    },
    "/api/admin/queue/bans": {
      "get": {
        "tags": ["Queue"],
        "summary": "Admin: list active queue bans",
        "responses": { "200": { "description": "Array of active ban rows" } }
      },
      "post": {
        "tags": ["Queue"],
        "summary": "Admin: create or update a queue ban",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["player_id"],
                "properties": {
                  "player_id": { "type": "integer" },
                  "duration_minutes": { "type": "integer", "description": "0 or null = permanent" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Ban saved" } }
      }
    },
    "/api/admin/queue/bans/{playerId}": {
      "delete": {
        "tags": ["Queue"],
        "summary": "Admin: remove a queue ban",
        "parameters": [{ "name": "playerId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Ban removed" } }
      }
    },
...
  }
}